+++ packages/patternfly-4/react-core/src/components/Alert/Alert.test.tsx
@@ -1,13 +1,13 @@
 import * as React from 'react';
-import { mount, shallow } from 'enzyme';
+import { mount } from 'enzyme';
 
 import { Alert, AlertVariant } from './Alert';
 import { AlertActionLink }  from './AlertActionLink';
 import { AlertActionCloseButton } from './AlertActionCloseButton';
 
 test('default Alert variant is info', () => {
-  const view = shallow(<Alert title="this is a test">Alert testing</Alert>);
-  expect(view.props().className).toMatch(/pf-m-info/);
+  const view = mount(<Alert title="this is a test">Alert testing</Alert>);
+  expect(view.find('Alert').childAt(0).prop('className')).toContain('pf-m-info');
 });
 
 Object.values(AlertVariant).forEach((variant) => {