+++ packages/patternfly-4/react-core/src/experimental/components/index.ts
@@ -1,3 +1,4 @@
 export * from './DataToolbar';
 export * from './Drawer';
 export * from './Spinner';
+export * from './OverflowMenu';