+++ packages/patternfly-4/react-integration/demo-app-ts/src/Demos.ts
@@ -362,6 +362,11 @@ export const Demos: DemoInterface[] = [
     componentType: Examples.OuiaDemo
   },
   {
+    id: 'overflow-menu-demo',
+    name: 'Overflow Menu Demo',
+    componentType: Examples.OverflowMenuDemo
+  },
+  {
     id: 'page-demo',
     name: 'Page Demo',
     componentType: Examples.PageDemo