+++ packages/patternfly-4/react-docs/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.13.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.13.28...@patternfly/react-docs@4.13.29) (2019-09-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.13.28](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.13.27...@patternfly/react-docs@4.13.28) (2019-09-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
+## [4.13.27](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.13.26...@patternfly/react-docs@4.13.27) (2019-09-26)
+
+**Note:** Version bump only for package @patternfly/react-docs
+
+
+
+
+
 ## [4.13.26](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-docs@4.13.25...@patternfly/react-docs@4.13.26) (2019-09-26)
 
 **Note:** Version bump only for package @patternfly/react-docs