+++ packages/patternfly-4/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.66](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.11.65...@patternfly/react-inline-edit-extension@2.11.66) (2019-09-26)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.11.65](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.11.64...@patternfly/react-inline-edit-extension@2.11.65) (2019-09-26)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.11.64](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.11.63...@patternfly/react-inline-edit-extension@2.11.64) (2019-09-26)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [2.11.63](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.11.62...@patternfly/react-inline-edit-extension@2.11.63) (2019-09-26)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
 ## [2.11.62](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@2.11.61...@patternfly/react-inline-edit-extension@2.11.62) (2019-09-26)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension